Step 1: Import and Concat Datasets
Note that the data must be downloaded in CSV format from the New Orleans
open data site and saved locally. Then, simply set the working directory
to the path on your machine to run the entire program.

This program imports all five years of calls data into R. Then, we add a
column to indicate the year of the call and concatenate the data into
one database. Finally, we output the resulting dataset as a .txt and
.RData format.