Step 2: Add more date/ time data
This program takes the output file from Step 1 and adds variables for
the month, day of the week, and hour of a call. Then, we output the data
again to .txt and .RData formats.

If you run Steps 1 and 2 concurrently, you need not save and re-load the
dataset. But, if there are any intermediate steps, it may be necessary.